pyproject.toml
[tool.poetry]
name = "odata1c-client"
version = "0.1.1"
description = "A Python client for interacting with 1C systems via OData v3, providing ORM-like functionality, data validation, and query construction."
authors = ["Vladislav Akashev <vlad.akashev.w@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "OData1C"}]


[tool.poetry.dependencies]
python = "^3.12"
pydantic = "2.10.3"
requests = "2.32.3"


[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
